+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 > Features:
+- [5d138d5] [Feature] Separate MCP server and client processes
 - [3d56ca2] [Feature] Add deepseek-chat (powered by DeepSeek-V3)
 - [64f29b8] [Feature] MCP tools support with OpenAI Assistant mode
 
 > Improvements:
 - [cccf3a3] [Improvement] Proper conversation handling when interrupted with tools
 - [5c93e35] [Improvement] Handle hallucinated tool calls (issue #19 by runekaagaard)
-- [62d5556] [Improvement] Create mcp_config from sample if missing
 
 
 ---